requirements.txt
django-bootstrap4
requests